Add strong typing for io/cvs with tests

+++ wwpdb/io/cvs/CvsUtility.py
@@ -24,36 +24,37 @@ import os
 import shutil
 import subprocess
 import tempfile
+from typing import List, Optional, Tuple
 
 
 class CvsWrapper:
     """Wrapper class for opertations on cvs repositories."""
 
-    def __init__(self, tmpPath="./"):
+    def __init__(self, tmpPath: str = "./") -> None:
         self.__tmpPath = tmpPath
         self.__logger = logging.getLogger("wwpdb.utils.rcsb.CvsWrapper")
         self.__logger.debug("Created instance of CvsWrapper")
         #
         # self.__debug = True
-        self.__repositoryHost = None
-        self.__repositoryPath = None
-        self.__cvsUser = None
-        self.__cvsPassword = None
-        self.__cvsRoot = None
+        self.__repositoryHost: Optional[str] = None
+        self.__repositoryPath: Optional[str] = None
+        self.__cvsUser: Optional[str] = None
+        self.__cvsPassword: Optional[str] = None
+        self.__cvsRoot: Optional[str] = None
         #
-        self.__wrkPath = None
+        self.__wrkPath: Optional[str] = None
         self.__cvsInfoFileName = "cvsInfo.txt"
         self.__cvsErrorFileName = "cvsError.txt"
 
-    def setRepositoryPath(self, host, path):
+    def setRepositoryPath(self, host: Optional[str], path: Optional[str]) -> None:
         self.__repositoryHost = host
         self.__repositoryPath = path
 
-    def setAuthInfo(self, user, password):
+    def setAuthInfo(self, user: Optional[str], password: Optional[str]) -> None:
         self.__cvsUser = user
         self.__cvsPassword = password
 
-    def getHistory(self, cvsPath):
+    def getHistory(self, cvsPath: str) -> str:
         text = ""
         cmd = self.__getHistoryCmd(cvsPath)
         if cmd is not None:
@@ -62,7 +63,7 @@ class CvsWrapper:
 
         return text
 
-    def getRevisionList(self, cvsPath):
+    def getRevisionList(self, cvsPath: str) -> List[Tuple[str, str, str]]:
         """Return a list of tuples containing the revision identifiers for the input file.
 
         Return data has the for [(RevId, A/M, timeStamp),...] where A=Added and M=Modified.
@@ -74,11 +75,13 @@ class CvsWrapper:
             revList = self.__extractRevisions()
         return revList
 
-    def cleanup(self):
+    def cleanup(self) -> None:  # Returns None on success or raises exception...
         """Cleanup temporary files and directories"""
+        if self.__wrkPath is None:
+            raise TypeError
         return shutil.rmtree(self.__wrkPath)
 
-    def checkOutFile(self, cvsPath, outPath, revId=None):
+    def checkOutFile(self, cvsPath: str, outPath: str, revId: Optional[str] = None) -> str:
         text = ""
         (pth, fn) = os.path.split(cvsPath)
         self.__logger.debug("Cvs directory %s   target file name %s", pth, fn)
@@ -89,20 +92,26 @@ class CvsWrapper:
                 text = self.__getErrorText()
         return text
 
-    def __getHistoryCmd(self, cvsPath):
+    def __getHistoryCmd(self, cvsPath: str) -> Optional[str]:
         if self.__wrkPath is None:
             self.__makeTempWorkingDir()
+        if self.__wrkPath is None:
+            raise ValueError
         outPath = os.path.join(self.__wrkPath, self.__cvsInfoFileName)
         errPath = os.path.join(self.__wrkPath, self.__cvsErrorFileName)
         if self.__setCvsRoot():
+            if self.__cvsRoot is None:
+                raise TypeError
             cmd = "cvs -d " + self.__cvsRoot + " history -a -x AM " + cvsPath + self.__getRedirect(fileNameOut=outPath, fileNameErr=errPath)
         else:
             cmd = None
         return cmd
 
-    def __getCheckOutCmd(self, cvsPath, outPath, revId=None):
+    def __getCheckOutCmd(self, cvsPath: str, outPath: str, revId: Optional[str] = None) -> Optional[str]:
         if self.__wrkPath is None:
             self.__makeTempWorkingDir()
+        if self.__wrkPath is None:
+            raise ValueError
         errPath = os.path.join(self.__wrkPath, self.__cvsErrorFileName)
         (pth, fn) = os.path.split(cvsPath)
         self.__logger.debug("CVS directory %s  target file name %s", pth, fn)
@@ -110,6 +119,8 @@ class CvsWrapper:
         #
         #
         if self.__setCvsRoot():
+            if self.__cvsRoot is None:
+                raise ValueError
             if revId is None:
                 rS = " "
                 cmd = (
@@ -151,7 +162,7 @@ class CvsWrapper:
             cmd = None
         return cmd
 
-    def __getRedirect(self, fileNameOut="myLog.log", fileNameErr="myLog.log", append=False):
+    def __getRedirect(self, fileNameOut: str = "myLog.log", fileNameErr: str = "myLog.log", append: bool = False) -> str:
         if append:
             if fileNameOut == fileNameErr:
                 oReDir = " >> " + fileNameOut + " 2>&1 "
@@ -164,7 +175,7 @@ class CvsWrapper:
 
         return oReDir
 
-    def __runCvsCommand(self, myCommand):
+    def __runCvsCommand(self, myCommand: str) -> bool:
         retcode = -100
         try:
             self.__logger.debug("Command: %s", myCommand)
@@ -188,18 +199,18 @@ class CvsWrapper:
             #    self.__lfh.write("+CvsWrapper(__runCvsCommand) Execution failed: %r\n" % e)
             return False
 
-    def __setCvsRoot(self):
+    def __setCvsRoot(self) -> bool:
         try:
-            self.__cvsRoot = ":pserver:" + self.__cvsUser + ":" + self.__cvsPassword + "@" + self.__repositoryHost + ":" + self.__repositoryPath
+            self.__cvsRoot = ":pserver:" + self.__cvsUser + ":" + self.__cvsPassword + "@" + self.__repositoryHost + ":" + self.__repositoryPath  # type: ignore
             return True
         except:  # noqa: E722 pylint: disable=bare-except
             return False
 
-    def __extractRevisions(self):
+    def __extractRevisions(self) -> List[Tuple[str, str, str]]:
         """Extract revisions details from the last history command."""
         revList = []
         try:
-            fName = os.path.join(self.__wrkPath, self.__cvsInfoFileName)
+            fName = os.path.join(self.__wrkPath, self.__cvsInfoFileName)  # type: ignore
             self.__logger.debug("Reading revisions from %r", fName)
             ifh = open(fName)
             for line in ifh:
@@ -216,10 +227,10 @@ class CvsWrapper:
 
         return revList
 
-    def __getOutputText(self):
+    def __getOutputText(self) -> str:
         text = ""
         try:
-            fPath = os.path.join(self.__wrkPath, self.__cvsInfoFileName)
+            fPath = os.path.join(self.__wrkPath, self.__cvsInfoFileName)  # type: ignore
             ifh = open(fPath)
             text = ifh.read()
         except:  # noqa: E722 pylint: disable=bare-except
@@ -227,10 +238,10 @@ class CvsWrapper:
 
         return text
 
-    def __getErrorText(self):
+    def __getErrorText(self) -> str:
         text = ""
         try:
-            fName = os.path.join(self.__wrkPath, self.__cvsErrorFileName)
+            fName = os.path.join(self.__wrkPath, self.__cvsErrorFileName)  # type: ignore
             ifh = open(fName)
             text = ifh.read()
         except:  # noqa: E722 pylint: disable=bare-except
@@ -238,7 +249,7 @@ class CvsWrapper:
 
         return text
 
-    def __makeTempWorkingDir(self):
+    def __makeTempWorkingDir(self) -> None:
         if self.__tmpPath is not None and os.path.isdir(self.__tmpPath):
             self.__wrkPath = tempfile.mkdtemp("tmpdir", "rcsbCVS", self.__tmpPath)
         else:
